refactor : UI 및 도메인 로직 분리를 위한 틀 생성
