+++ src/main/kotlin/lotto/controller/LottoController.kt
@@ -1,0 +1,5 @@
+package lotto.controller
+
+class LottoController {
+    // TODO 게임플레이를 관리한다.
+}
