+++ src/main/kotlin/lotto/domain/LottoGame.kt
@@ -1,0 +1,5 @@
+package lotto.domain
+
+class LottoGame {
+    // TODO 로또게임에 필요한 데이터를 정의한다.
+}
