+++ src/main/kotlin/lotto/domain/model/Prize.kt
@@ -1,6 +1,6 @@
-package lotto
+package lotto.domain.model
 
-import lotto.Lotto.Companion.ERROR_MESSAGE
+import lotto.domain.model.Lotto.Companion.ERROR_MESSAGE
 
 enum class Prize {
     CORRECT_THREE,
@@ -15,7 +15,7 @@ enum class Prize {
                 4 -> CORRECT_FOUR
                 5 -> CORRECT_FIVE
                 6 -> CORRECT_SIX
-                else -> throw IllegalStateException("$ERROR_MESSAGE 당첨 갯수가 잘못됬습니다.")
+                else -> throw IllegalStateException("$ERROR_MESSAGE 당첨 갯수가 잘못됐습니다.")
             }
         }
     }
