+++ src/main/kotlin/lotto/view/InputView.kt
@@ -1,0 +1,6 @@
+package lotto.view
+
+class InputView {
+    // TODO 유저의 입력을 받아온다.
+
+}
