+++ src/main/kotlin/lotto/view/OutputView.kt
@@ -1,0 +1,5 @@
+package lotto.view
+
+class OutputView {
+    // TODO 출력을 담당하는 뷰
+}
